src/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Markdown==3.8
openai==1.82.0
pydantic==2.11.5
pydantic_core==2.33.2
pytube @ git+https://github.com/pytube/pytube.git@a32fff39058a6f7e5e59ecd06a7467b71197ce35
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
youtube-transcript-api==1.0.3
